feat(graph-web): demo project
