+++ testProject/.classpath
@@ -2,7 +2,24 @@
 <classpath>
 	<classpathentry kind="src" path="javasource"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.5.0.25874/runtime/bundles/com.mendix.json.jar"/><classpathentry kind="lib" path="C:/Program Files/Mendix/9.5.0.25874/runtime/bundles/com.mendix.logging-api.jar"/><classpathentry kind="lib" path="C:/Program Files/Mendix/9.5.0.25874/runtime/bundles/com.mendix.m2ee-api.jar"/><classpathentry kind="lib" path="C:/Program Files/Mendix/9.5.0.25874/runtime/bundles/com.mendix.public-api.jar"/><classpathentry kind="lib" path="C:/Program Files/Mendix/9.5.0.25874/runtime/bundles/javax.servlet-api.servlet.jar"/><classpathentry kind="lib" path="C:/Program Files/Mendix/9.5.0.25874/runtime/bundles/javax.websocket-api.websocket.jar"/>
-	<classpathentry kind="lib" path="userlib/checker-qual-3.8.0.jar"/><classpathentry kind="lib" path="userlib/commons-io-2.8.0.jar"/><classpathentry kind="lib" path="userlib/commons-lang3-3.11.jar"/><classpathentry kind="lib" path="userlib/commons-logging-1.2.jar"/><classpathentry kind="lib" path="userlib/error_prone_annotations-2.5.1.jar"/><classpathentry kind="lib" path="userlib/failureaccess-1.0.1.jar"/><classpathentry kind="lib" path="userlib/fontbox-2.0.24.jar"/><classpathentry kind="lib" path="userlib/guava-30.1.1-jre.jar"/><classpathentry kind="lib" path="userlib/j2objc-annotations-1.3.jar"/><classpathentry kind="lib" path="userlib/jsr305-3.0.2.jar"/><classpathentry kind="lib" path="userlib/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar"/><classpathentry kind="lib" path="userlib/owasp-java-html-sanitizer-20200713.1.jar"/><classpathentry kind="lib" path="userlib/pdfbox-2.0.24.jar"/>
+	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.6.2.30227/runtime/bundles/com.mendix.json.jar"/>
+	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.6.2.30227/runtime/bundles/com.mendix.logging-api.jar"/>
+	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.6.2.30227/runtime/bundles/com.mendix.m2ee-api.jar"/>
+	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.6.2.30227/runtime/bundles/com.mendix.public-api.jar"/>
+	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.6.2.30227/runtime/bundles/javax.servlet-api.servlet.jar"/>
+	<classpathentry kind="lib" path="C:/Program Files/Mendix/9.6.2.30227/runtime/bundles/javax.websocket-api.websocket.jar"/>
+	<classpathentry kind="lib" path="userlib/checker-qual-3.8.0.jar"/>
+	<classpathentry kind="lib" path="userlib/commons-io-2.8.0.jar"/>
+	<classpathentry kind="lib" path="userlib/commons-lang3-3.11.jar"/>
+	<classpathentry kind="lib" path="userlib/commons-logging-1.2.jar"/>
+	<classpathentry kind="lib" path="userlib/error_prone_annotations-2.5.1.jar"/>
+	<classpathentry kind="lib" path="userlib/failureaccess-1.0.1.jar"/>
+	<classpathentry kind="lib" path="userlib/fontbox-2.0.24.jar"/>
+	<classpathentry kind="lib" path="userlib/guava-30.1.1-jre.jar"/>
+	<classpathentry kind="lib" path="userlib/j2objc-annotations-1.3.jar"/>
+	<classpathentry kind="lib" path="userlib/jsr305-3.0.2.jar"/>
+	<classpathentry kind="lib" path="userlib/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar"/>
+	<classpathentry kind="lib" path="userlib/owasp-java-html-sanitizer-20200713.1.jar"/>
+	<classpathentry kind="lib" path="userlib/pdfbox-2.0.24.jar"/>
 	<classpathentry kind="output" path="deployment/run/bin"/>
 </classpath>
