+++ testProject/javasource/system/proxies/ContextType.java
@@ -1,0 +1,36 @@
+// This file was generated by Mendix Studio Pro.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package system.proxies;
+
+/**
+ * The type of context to use for executing queued tasks.
+ */
+public enum ContextType
+{
+	System(new java.lang.String[][] { new java.lang.String[] { "en_US", "System" } }),
+	User(new java.lang.String[][] { new java.lang.String[] { "en_US", "User" } }),
+	Anonymous(new java.lang.String[][] { new java.lang.String[] { "en_US", "Anonymous" } });
+
+	private java.util.Map<java.lang.String, java.lang.String> captions;
+
+	private ContextType(java.lang.String[][] captionStrings)
+	{
+		this.captions = new java.util.HashMap<java.lang.String, java.lang.String>();
+		for (java.lang.String[] captionString : captionStrings)
+			captions.put(captionString[0], captionString[1]);
+	}
+
+	public java.lang.String getCaption(java.lang.String languageCode)
+	{
+		if (captions.containsKey(languageCode))
+			return captions.get(languageCode);
+		return captions.get("en_US");
+	}
+
+	public java.lang.String getCaption()
+	{
+		return captions.get("en_US");
+	}
+}
