--- testProject/javasource/system/proxies/WorkflowSystemTask.java
@@ -1,123 +1,0 @@
-// This file was generated by Mendix Studio Pro.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package system.proxies;
-
-public class WorkflowSystemTask extends system.proxies.WorkflowTaskInstance
-{
-	/**
-	 * Internal name of this entity
-	 */
-	public static final java.lang.String entityName = "System.WorkflowSystemTask";
-
-	/**
-	 * Enum describing members of this entity
-	 */
-	public enum MemberNames
-	{
-		Outcome("Outcome"),
-		StartTime("StartTime"),
-		EndTime("EndTime"),
-		State("State"),
-		Reason("Reason"),
-		TaskInstance_WorkflowInstance("System.TaskInstance_WorkflowInstance"),
-		TaskInstance_TaskDefinition("System.TaskInstance_TaskDefinition");
-
-		private java.lang.String metaName;
-
-		MemberNames(java.lang.String s)
-		{
-			metaName = s;
-		}
-
-		@java.lang.Override
-		public java.lang.String toString()
-		{
-			return metaName;
-		}
-	}
-
-	public WorkflowSystemTask(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		this(context, com.mendix.core.Core.instantiate(context, "System.WorkflowSystemTask"));
-	}
-
-	protected WorkflowSystemTask(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject workflowSystemTaskMendixObject)
-	{
-		super(context, workflowSystemTaskMendixObject);
-		if (!com.mendix.core.Core.isSubClassOf("System.WorkflowSystemTask", workflowSystemTaskMendixObject.getType()))
-			throw new java.lang.IllegalArgumentException("The given object is not a System.WorkflowSystemTask");
-	}
-
-	/**
-	 * @deprecated Use 'WorkflowSystemTask.load(IContext, IMendixIdentifier)' instead.
-	 */
-	@java.lang.Deprecated
-	public static system.proxies.WorkflowSystemTask initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		return system.proxies.WorkflowSystemTask.load(context, mendixIdentifier);
-	}
-
-	/**
-	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
-	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
-	 */
-	public static system.proxies.WorkflowSystemTask initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
-	{
-		return new system.proxies.WorkflowSystemTask(context, mendixObject);
-	}
-
-	public static system.proxies.WorkflowSystemTask load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
-		return system.proxies.WorkflowSystemTask.initialize(context, mendixObject);
-	}
-
-	public static java.util.List<system.proxies.WorkflowSystemTask> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
-	{
-		java.util.List<system.proxies.WorkflowSystemTask> result = new java.util.ArrayList<system.proxies.WorkflowSystemTask>();
-		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//System.WorkflowSystemTask" + xpathConstraint))
-			result.add(system.proxies.WorkflowSystemTask.initialize(context, obj));
-		return result;
-	}
-
-	@java.lang.Override
-	public boolean equals(Object obj)
-	{
-		if (obj == this)
-			return true;
-
-		if (obj != null && getClass().equals(obj.getClass()))
-		{
-			final system.proxies.WorkflowSystemTask that = (system.proxies.WorkflowSystemTask) obj;
-			return getMendixObject().equals(that.getMendixObject());
-		}
-		return false;
-	}
-
-	@java.lang.Override
-	public int hashCode()
-	{
-		return getMendixObject().hashCode();
-	}
-
-	/**
-	 * @return String name of this class
-	 */
-	public static java.lang.String getType()
-	{
-		return "System.WorkflowSystemTask";
-	}
-
-	/**
-	 * @return String GUID from this object, format: ID_0000000000
-	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
-	 */
-	@java.lang.Override
-	@java.lang.Deprecated
-	public java.lang.String getGUID()
-	{
-		return "ID_" + getMendixObject().getId().toLong();
-	}
-}
