--- testProject/javasource/system/proxies/WorkflowTaskDefinition.java
@@ -1,288 +1,0 @@
-// This file was generated by Mendix Studio Pro.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package system.proxies;
-
-public class WorkflowTaskDefinition
-{
-	private final com.mendix.systemwideinterfaces.core.IMendixObject workflowTaskDefinitionMendixObject;
-
-	private final com.mendix.systemwideinterfaces.core.IContext context;
-
-	/**
-	 * Internal name of this entity
-	 */
-	public static final java.lang.String entityName = "System.WorkflowTaskDefinition";
-
-	/**
-	 * Enum describing members of this entity
-	 */
-	public enum MemberNames
-	{
-		Name("Name"),
-		IsObsolete("IsObsolete"),
-		TaskDefinition_WorkflowDefinition("System.TaskDefinition_WorkflowDefinition");
-
-		private java.lang.String metaName;
-
-		MemberNames(java.lang.String s)
-		{
-			metaName = s;
-		}
-
-		@java.lang.Override
-		public java.lang.String toString()
-		{
-			return metaName;
-		}
-	}
-
-	public WorkflowTaskDefinition(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		this(context, com.mendix.core.Core.instantiate(context, "System.WorkflowTaskDefinition"));
-	}
-
-	protected WorkflowTaskDefinition(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject workflowTaskDefinitionMendixObject)
-	{
-		if (workflowTaskDefinitionMendixObject == null)
-			throw new java.lang.IllegalArgumentException("The given object cannot be null.");
-		if (!com.mendix.core.Core.isSubClassOf("System.WorkflowTaskDefinition", workflowTaskDefinitionMendixObject.getType()))
-			throw new java.lang.IllegalArgumentException("The given object is not a System.WorkflowTaskDefinition");
-
-		this.workflowTaskDefinitionMendixObject = workflowTaskDefinitionMendixObject;
-		this.context = context;
-	}
-
-	/**
-	 * @deprecated Use 'WorkflowTaskDefinition.load(IContext, IMendixIdentifier)' instead.
-	 */
-	@java.lang.Deprecated
-	public static system.proxies.WorkflowTaskDefinition initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		return system.proxies.WorkflowTaskDefinition.load(context, mendixIdentifier);
-	}
-
-	/**
-	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
-	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
-	 */
-	public static system.proxies.WorkflowTaskDefinition initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
-	{
-		return new system.proxies.WorkflowTaskDefinition(context, mendixObject);
-	}
-
-	public static system.proxies.WorkflowTaskDefinition load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
-		return system.proxies.WorkflowTaskDefinition.initialize(context, mendixObject);
-	}
-
-	public static java.util.List<system.proxies.WorkflowTaskDefinition> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
-	{
-		java.util.List<system.proxies.WorkflowTaskDefinition> result = new java.util.ArrayList<system.proxies.WorkflowTaskDefinition>();
-		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//System.WorkflowTaskDefinition" + xpathConstraint))
-			result.add(system.proxies.WorkflowTaskDefinition.initialize(context, obj));
-		return result;
-	}
-
-	/**
-	 * Commit the changes made on this proxy object.
-	 */
-	public final void commit() throws com.mendix.core.CoreException
-	{
-		com.mendix.core.Core.commit(context, getMendixObject());
-	}
-
-	/**
-	 * Commit the changes made on this proxy object using the specified context.
-	 */
-	public final void commit(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
-	{
-		com.mendix.core.Core.commit(context, getMendixObject());
-	}
-
-	/**
-	 * Delete the object.
-	 */
-	public final void delete()
-	{
-		com.mendix.core.Core.delete(context, getMendixObject());
-	}
-
-	/**
-	 * Delete the object using the specified context.
-	 */
-	public final void delete(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		com.mendix.core.Core.delete(context, getMendixObject());
-	}
-	/**
-	 * @return value of Name
-	 */
-	public final java.lang.String getName()
-	{
-		return getName(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of Name
-	 */
-	public final java.lang.String getName(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.String) getMendixObject().getValue(context, MemberNames.Name.toString());
-	}
-
-	/**
-	 * Set value of Name
-	 * @param name
-	 */
-	public final void setName(java.lang.String name)
-	{
-		setName(getContext(), name);
-	}
-
-	/**
-	 * Set value of Name
-	 * @param context
-	 * @param name
-	 */
-	public final void setName(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String name)
-	{
-		getMendixObject().setValue(context, MemberNames.Name.toString(), name);
-	}
-
-	/**
-	 * @return value of IsObsolete
-	 */
-	public final java.lang.Boolean getIsObsolete()
-	{
-		return getIsObsolete(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of IsObsolete
-	 */
-	public final java.lang.Boolean getIsObsolete(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Boolean) getMendixObject().getValue(context, MemberNames.IsObsolete.toString());
-	}
-
-	/**
-	 * Set value of IsObsolete
-	 * @param isobsolete
-	 */
-	public final void setIsObsolete(java.lang.Boolean isobsolete)
-	{
-		setIsObsolete(getContext(), isobsolete);
-	}
-
-	/**
-	 * Set value of IsObsolete
-	 * @param context
-	 * @param isobsolete
-	 */
-	public final void setIsObsolete(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Boolean isobsolete)
-	{
-		getMendixObject().setValue(context, MemberNames.IsObsolete.toString(), isobsolete);
-	}
-
-	/**
-	 * @return value of TaskDefinition_WorkflowDefinition
-	 */
-	public final system.proxies.WorkflowDefinition getTaskDefinition_WorkflowDefinition() throws com.mendix.core.CoreException
-	{
-		return getTaskDefinition_WorkflowDefinition(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of TaskDefinition_WorkflowDefinition
-	 */
-	public final system.proxies.WorkflowDefinition getTaskDefinition_WorkflowDefinition(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
-	{
-		system.proxies.WorkflowDefinition result = null;
-		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.TaskDefinition_WorkflowDefinition.toString());
-		if (identifier != null)
-			result = system.proxies.WorkflowDefinition.load(context, identifier);
-		return result;
-	}
-
-	/**
-	 * Set value of TaskDefinition_WorkflowDefinition
-	 * @param taskdefinition_workflowdefinition
-	 */
-	public final void setTaskDefinition_WorkflowDefinition(system.proxies.WorkflowDefinition taskdefinition_workflowdefinition)
-	{
-		setTaskDefinition_WorkflowDefinition(getContext(), taskdefinition_workflowdefinition);
-	}
-
-	/**
-	 * Set value of TaskDefinition_WorkflowDefinition
-	 * @param context
-	 * @param taskdefinition_workflowdefinition
-	 */
-	public final void setTaskDefinition_WorkflowDefinition(com.mendix.systemwideinterfaces.core.IContext context, system.proxies.WorkflowDefinition taskdefinition_workflowdefinition)
-	{
-		if (taskdefinition_workflowdefinition == null)
-			getMendixObject().setValue(context, MemberNames.TaskDefinition_WorkflowDefinition.toString(), null);
-		else
-			getMendixObject().setValue(context, MemberNames.TaskDefinition_WorkflowDefinition.toString(), taskdefinition_workflowdefinition.getMendixObject().getId());
-	}
-
-	/**
-	 * @return the IMendixObject instance of this proxy for use in the Core interface.
-	 */
-	public final com.mendix.systemwideinterfaces.core.IMendixObject getMendixObject()
-	{
-		return workflowTaskDefinitionMendixObject;
-	}
-
-	/**
-	 * @return the IContext instance of this proxy, or null if no IContext instance was specified at initialization.
-	 */
-	public final com.mendix.systemwideinterfaces.core.IContext getContext()
-	{
-		return context;
-	}
-
-	@java.lang.Override
-	public boolean equals(Object obj)
-	{
-		if (obj == this)
-			return true;
-
-		if (obj != null && getClass().equals(obj.getClass()))
-		{
-			final system.proxies.WorkflowTaskDefinition that = (system.proxies.WorkflowTaskDefinition) obj;
-			return getMendixObject().equals(that.getMendixObject());
-		}
-		return false;
-	}
-
-	@java.lang.Override
-	public int hashCode()
-	{
-		return getMendixObject().hashCode();
-	}
-
-	/**
-	 * @return String name of this class
-	 */
-	public static java.lang.String getType()
-	{
-		return "System.WorkflowTaskDefinition";
-	}
-
-	/**
-	 * @return String GUID from this object, format: ID_0000000000
-	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
-	 */
-	@java.lang.Deprecated
-	public java.lang.String getGUID()
-	{
-		return "ID_" + getMendixObject().getId().toLong();
-	}
-}
